xf86drm: Add drmGetDeviceFromMajorMinor
This adds a function to get a drmDevicePtr from a major+minor number of
a device. This is useful for Wayland that uses these to identify devices
over the protocol.

This is done by taking the implementation of drmGetDevice2, and removing
the call to fstat to find the major+minor.

Signed-off-by: Scott Anderson <scott.anderson@collabora.com>