xf86drm: add support for populating drm formats
This change lets libdrm take care of allocation and deallocation of
DRM plane formats on behalf of clients. Weston, xserver and others
can use this functionality instead of querying and parsing the blob.

This change makes two assumptions:

  1. drm kernel blobs are expected to be backwards compatible,
     so newer kernels should not break an older userspace
  2. newer versions of libdrm must not break apps linked against
     an older version of libdrm

Assumption (1) is achieved in the kernel and in libdrm. In the kernel,
extending KMS property blobs (drm_format_modifier_blob for IN_FORMATS)
with only additional members beyond the currently-defined fields. In
libdrm, by accessing those fields with an offset reflecting the acutal
layout of the blob.

Assumption (2) is achieved in libdrm. Specifically introducing a new
set of data types and symbols for supporting new blobs and leaving
unaltered old data types and symbols for backwards compatibility.

As an example, support for DRM/KMS IN_FORMATS blob version 2 could
be achieved introducing the following new set of symbols:

   drmModePopulateFormats2
   drmModeAllocFormats2
   drmModeFreeFormats2

and a new data type:

   struct _drmModeFormats2

Signed-off-by: Luigi Santivetti <luigi.santivetti@imgtec.com>