drmModeAtomicCommit: stable sort for requests
A drmModeAtomicReq can contain multiple values for the same property,
with later cursor values overriding earlier. For example, this should
result in the property being set to 0x4:
  drmModeAtomicReqPtr req = drmModeAtomicAlloc();
  drmModeAtomicAddProperty(req, 0x1, 0x2, 0x2);
  drmModeAtomicAddProperty(req, 0x1, 0x2, 0x4);
  drmModeAtomicCommit(fd, req, 0, NULL);

drmModeAtomicCommit already sorts the property set by object and
property ID. We were relying on qsort to also sort by cursor - i.e.
pointer value - when object and property ID are equal, however whilst
glibc does this, the sort order is explicitly undefined when the
comparator is equal.

Add a pointer comparator to the sort operator to ensure that the sorted
list is chronologically ordered for the same object and property.

Fixes: #46

Signed-off-by: Daniel Stone <daniels@collabora.com>