test/amdgpu: modify secure bounce test
The original test steps are:
1. Use memset to write pattern to encrypted buffer A;
2. Sdma clear copy from encrypted buffer A to regular buffer B;
3. Move A to GTT domain;
4. Sdma clear copy back to encrypted buffer A.
The TMZ data can be only decrypted successfully in TMZ mode.
For step2, it's an invalid use case for secure/clear copy if the
target buffer is not TMZ buffer.
For step4, it's also an invalid use case to clear copy from the
clear buffer to TMZ buffer.

The new steps are following:
1. allocate regular buffer A and D, encrypted buffer B and C;
2. Fill constant pattern into A;
3. sdma secure copy from A to B;
4. Move B to GTT domain,
   Check whether the encrypted data of B is same to A;
5. sdma secure copy from B to C;
6. Move C to GTT domain,
   Check whether the encrypted data of C is same to A;
7. sdma clear copy from B to C;
   Check whether the encrypted data of C is different to A.

Signed-off-by: Aaron Liu <aaron.liu@amd.com>