test/amdgpu: modify secure bounce test(V2)
The original test steps are:
1. Use memset to write pattern to alice (encrypted buffer);
2. Sdma clear copy from alice to bob(regular buffer);
3. Move alice to GTT domain;
4. Sdma clear copy back to alice.
The TMZ data can be only decrypted successfully in TMZ mode.
For step2, it's an invalid use case for secure/clear copy if the
target buffer is not TMZ buffer.
For step4, it's also an invalid use case to clear copy from the
clear buffer to TMZ buffer.

The new steps are following:
1. allocate regular buffer alice and dave, encrypted buffer
bob and charlie;
2. Fill constant pattern into alice;
3. sdma secure copy from alice to bob;
4. Move bob to GTT domain,
   Check whether the encrypted data of bob is same to alice;
5. sdma secure copy from bob to charlie;
6. Move charlie to GTT domain,
   Check whether the encrypted data of charlie is same to alice;
7. sdma clear copy from charlie to dave;
   Check whether the encrypted data of dave is different to alice.

v2: https://gitlab.freedesktop.org/mesa/drm/-/merge_requests/183

Signed-off-by: Aaron Liu <aaron.liu@amd.com>